+++ src/dir_to_undir_pp.cpp
@@ -80,17 +80,17 @@ int dir_to_undir::parse_params() {
 			return 1;
 		}
 		else if (param == "-n") {
-			m_n = static_cast<uint32_t>(atoi(m_argv[i + 1]));
+			m_n = static_cast<uint64_t>(atoi(m_argv[i + 1]));
 			m_has_n = true;
 			++i;
 		}
 		else if (param == "-T") {
-			m_T = static_cast<uint32_t>(atoi(m_argv[i + 1]));
+			m_T = static_cast<uint64_t>(atoi(m_argv[i + 1]));
 			m_has_T = true;
 			++i;
 		}
 		else if (param == "-C") {
-			m_C = static_cast<uint32_t>(atoi(m_argv[i + 1]));
+			m_C = static_cast<uint64_t>(atoi(m_argv[i + 1]));
 			++i;
 		}
 		else if (param == "-mode") {
