+++ src/generate.cpp
@@ -49,7 +49,7 @@ edge gimme_edge(const T& t) {
 	return it.get_edge();
 }
 
-void output_execution_time_trees(double total_ms, uint32_t n, uint32_t N, uint32_t R) {
+void output_execution_time_trees(double total_ms, uint64_t n, uint64_t N, uint64_t R) {
 	cout << "n= " << n << endl;
 	cout << "N= " << N << endl;
 	cout << "R= " << R << endl;
@@ -59,12 +59,12 @@ void output_execution_time_trees(double total_ms, uint32_t n, uint32_t N, uint32
 }
 
 template<class tree_type, class GEN>
-void profile_exhaustive_trees(uint32_t n, uint32_t N, uint32_t R) {
+void profile_exhaustive_trees(uint64_t n, uint64_t N, uint64_t R) {
 	double total = 0.0;
 
-	for (uint32_t r = 0; r < R; ++r) {
+	for (uint64_t r = 0; r < R; ++r) {
 		GEN Gen(n);
-		for (uint32_t i = 0; i < N and not Gen.end(); ++i) {
+		for (uint64_t i = 0; i < N and not Gen.end(); ++i) {
 			const auto begin = profiling::now();
 			tree_type tree = Gen.get_tree();
 			Gen.next();
@@ -83,16 +83,16 @@ void profile_exhaustive_trees(uint32_t n, uint32_t N, uint32_t R) {
 }
 
 template<class tree_type, class tree_gen_type>
-void profile_random_trees(uint32_t n, uint32_t N, uint32_t R) {
+void profile_random_trees(uint64_t n, uint64_t N, uint64_t R) {
 	double total = 0.0;
 
-	for (uint32_t r = 0; r < R; ++r) {
+	for (uint64_t r = 0; r < R; ++r) {
 		tree_gen_type Gen(n, 1234);
 		Gen.set_calculate_size_subtrees(false);
 		Gen.set_normalise_tree(false);
 		Gen.set_calculate_tree_type(false);
 
-		for (uint32_t i = 0; i < N; ++i) {
+		for (uint64_t i = 0; i < N; ++i) {
 			const auto begin = profiling::now();
 			tree_type tree = Gen.get_tree();
 			const auto end = profiling::now();
@@ -118,9 +118,9 @@ void generate_trees(int argc, char *argv[]) {
 	}
 
 	const string& what = parser.get_gen_class();
-	const uint32_t n = parser.get_n();
-	const uint32_t N = parser.get_N();
-	const uint32_t R = parser.get_R();
+	const uint64_t n = parser.get_n();
+	const uint64_t N = parser.get_N();
+	const uint64_t R = parser.get_R();
 
 	if (what == "all_lab_free") {
 		generate::profile_exhaustive_trees<free_tree, all_lab_free_trees>(n, N, R);
@@ -157,7 +157,7 @@ void generate_trees(int argc, char *argv[]) {
 namespace generate {
 
 void output_execution_time_arrangements
-(double total_ms, uint32_t n, uint32_t R, uint32_t T, uint32_t N)
+(double total_ms, uint64_t n, uint64_t R, uint64_t T, uint64_t N)
 {
 	cout << "n= " << n << endl;
 	cout << "R= " << R << endl;
@@ -179,11 +179,11 @@ void output_execution_time_arrangements
 
 template<class tree_type, class tree_randgen_type, class arr_gen_type>
 void profile_exhaustive_arrangements
-(uint32_t n, uint32_t R, uint32_t T, uint32_t N)
+(uint64_t n, uint64_t R, uint64_t T, uint64_t N)
 {
 	double total = 0.0;
 
-	for (uint32_t r = 0; r < R; ++r) {
+	for (uint64_t r = 0; r < R; ++r) {
 		tree_randgen_type TreeGen(n);
 		for (size_t i = 0; i < T; ++i) {
 			const tree_type randtree = TreeGen.get_tree();
@@ -207,11 +207,11 @@ void profile_exhaustive_arrangements
 
 template<class tree_type, class tree_randgen_type, class arr_gen_type>
 void profile_random_arrangements
-(uint32_t n, uint32_t R, uint32_t T, uint32_t N)
+(uint64_t n, uint64_t R, uint64_t T, uint64_t N)
 {
 	double total = 0.0;
 
-	for (uint32_t r = 0; r < R; ++r) {
+	for (uint64_t r = 0; r < R; ++r) {
 		tree_randgen_type TreeGen(n);
 
 		for (size_t i = 0; i < T; ++i) {
@@ -241,10 +241,10 @@ void generate_arrangements(int argc, char *argv[]) {
 	}
 
 	const string& what = parser.get_gen_class();
-	const uint32_t n = parser.get_n();
-	const uint32_t R = parser.get_R();
-	const uint32_t T = parser.get_T();
-	const uint32_t N = parser.get_N();
+	const uint64_t n = parser.get_n();
+	const uint64_t R = parser.get_R();
+	const uint64_t T = parser.get_T();
+	const uint64_t N = parser.get_N();
 
 	if (what == "all_arrangements") {
 		generate::profile_exhaustive_arrangements
