+++ src/generate_trees_pp.cpp
@@ -82,17 +82,17 @@ int generate_trees_pp::parse_params() noexcept {
 			return 1;
 		}
 		else if (param == "-n") {
-			m_n = static_cast<uint32_t>(atoi(m_argv[i + 1]));
+			m_n = static_cast<uint64_t>(atoi(m_argv[i + 1]));
 			m_has_n = true;
 			++i;
 		}
 		else if (param == "-N") {
-			m_N = static_cast<uint32_t>(atoi(m_argv[i + 1]));
+			m_N = static_cast<uint64_t>(atoi(m_argv[i + 1]));
 			m_has_N = true;
 			++i;
 		}
 		else if (param == "-R") {
-			m_R = static_cast<uint32_t>(atoi(m_argv[i + 1]));
+			m_R = static_cast<uint64_t>(atoi(m_argv[i + 1]));
 			m_has_R = true;
 			++i;
 		}
