+++ src/linarr_C_pp.hpp
@@ -37,9 +37,9 @@ public:
 	~linarr_C_pp();
 
 	const std::string& get_algo() const { return m_gen_algo; }
-	constexpr uint32_t get_n() const { return m_n; }
-	constexpr uint32_t get_N() const { return m_N; }
-	constexpr uint32_t get_T() const { return m_T; }
+	constexpr uint64_t get_n() const { return m_n; }
+	constexpr uint64_t get_N() const { return m_N; }
+	constexpr uint64_t get_T() const { return m_T; }
 
 	void print_usage() const;
 
@@ -57,15 +57,15 @@ private:
 	std::string m_gen_algo = "none";
 
 	// number of vertices
-	uint32_t m_n = 0;
+	uint64_t m_n = 0;
 	bool m_has_n = false;
 
 	// number of trees to generate
-	uint32_t m_N = 0;
+	uint64_t m_N = 0;
 	bool m_has_N = false;
 
 	// number of replicas (times to repeat the same execution)
-	uint32_t m_T = 0;
+	uint64_t m_T = 0;
 	bool m_has_T = false;
 
 	const std::set<std::string> m_allowed_algorithms =
